Reuse code in tests
