--- source/tests/d3d9/logger.h
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2017- Hourglass Resurrection Team
- * Hourglass Resurrection is licensed under GPL v2.
- * Refer to the file COPYING.txt in the project root.
- */
-
-#pragma once
-
-namespace Logger
-{
-    void Write(const std::wstring& line);
-    void WriteLine(const std::wstring& line);
-}
