+++ source/tests/shared/logger.cpp
